Adding logical layer and editing put
